requirements.txt
diffusers
einops
modelscope
numpy
oss2
Pillow
PyYAML
safetensors
tqdm
imgviz
transformers
tensorboard
accelerate
opencv-python
imgviz
oss2
openai-clip
scikit-learn